% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RWAshp}
\alias{RWAshp}
\title{First-level administrative units for Rwanda}
\format{A shapefile of first-level administrative units for Rwanda}
\source{
\url{http://gadm.org/}
}
\usage{
RWAshp
}
\description{
First-level administrative units for Rwanda
}
\keyword{datasets}
